Various bug fixes (28 March 2013)
	results_analysis_matlab/routines2D/creatPics.m
	results_analysis_matlab/routines2D/utilities.getProgramPath.m
Made the two scripts to be more platform-independent. The booleans ispc
and isunix tell MatLab whether the platform is Windows or Unix-based
(e.g. Linux, Mac) and the corresponding command is called. All other
minor changes stem from this.

	src/simulator/Simulator.java
Removed references to writeInvReport(), a reduced report-writing scheme
for invasion/competition simulations which is now redundant

	src/simulator/World.java
Improved error reporting

	src/simulator/ActiveAgent.java
	src/simulator/LocatedAgent.java
Upon cell division, cells now update their growth and volume rates
(approximately half of what they were before). This has no effect on
simulations, but improves reporting

	src/simulator/Agent.java
	src/utils/ExtraMath.java
Completion of the update to make _genealogy a long variable

	build.xml
The ant file to compile binaries if the user is not using Eclipse (or
has iDynoMiCS stored remotely on a server which does not have Eclipse)