Random seed not specified in protocol
Previously, if there was no random seed specified in the protocol file
iDynoMiCS would use zero instead. This new method picks a random seed
that is the product of Math.random() and the system time in
milliseconds.

The random seed is then included in the log file for later reference.