Improved reporting of netGrowthRate
Improved reporting of netGrowthRate when autocatalytic reactions are
involved: we previously used the same estimator as non-autocatalytic
reactions, but this gives the growth rate at the start of the agent
timestep. The new method uses the average growth rate over the agent
time step.

Note that if (agentTimeStep < outputPeriod) then the reported growth
rate will be that of the final agent timestep in the output period and
not the overall growth rate of the output period.